Deduplicate board layout
