+++ src/routes/board/[boardName]/+page.svelte
@@ -1,5 +1,4 @@
 <script lang="ts">
-	import HelpText from "$lib/components/HelpText.svelte";
 	import PostList from "$lib/components/PostList.svelte";
 
 	let { data } = $props();
@@ -48,16 +47,4 @@
 	]);
 </script>
 
-<div
-	class="container mx-auto flex flex-col gap-6 px-4 py-6 sm:px-6 md:flex-row md:items-baseline lg:px-8"
->
-	<HelpText>
-		{#snippet title()}
-			{data.board.name}
-		{/snippet}
-		{#snippet body()}
-			{data.board.description}
-		{/snippet}
-	</HelpText>
-	<PostList {posts} boardName={data.board.name} />
-</div>
+<PostList {posts} boardName={data.board.name} />
